add person grid component and restructure to have a people folder for personcard and person grid components
