+++ components/people/index.js
@@ -1,0 +1,2 @@
+export * from './person-card'
+export * from './person-grid'
